Multi-network gas estimates (#456)
* Add GasService and providers

* Use GasePriceService over util fn

* Remove util fn
